# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "HT Lotto API",
        "description": "REST API for HT ភ្នាក់ Lottery Agent App. Roles: admin | master | staff.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API v1"
        }
    ],
    "paths": {
...
        "/bets": {
            "get": {
                "tags": [
                    "Bets"
                ],
                "summary": "List today's tickets — staff/master: own only; admin: all",
                "operationId": "a7d49cc1d5a255786e371ca42632a59e",
                "parameters": [
                    {
                        "name": "date",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2026-05-02"
                        }
                    },
                    {
                        "name": "session",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "morning",
                                "noon",
                                "evening"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ticket list",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Ticket"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Bets"
                ],
                "summary": "Submit a new ticket with bets — all roles",
                "operationId": "92f4acd29afae2866fb0dbfcf337b1f7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "session",
                                    "bet_date",
                                    "bets"
                                ],
                                "properties": {
                                    "session": {
                                        "type": "string",
                                        "enum": [
                                            "morning",
                                            "noon",
                                            "evening"
                                        ]
                                    },
                                    "bet_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-05-02"
                                    },
                                    "bets": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "bet_type": {
                                                    "type": "string",
                                                    "enum": [
                                                        "ABCD",
                                                        "LO"
                                                    ]
                                                },
                                                "letter": {
                                                    "type": "string",
                                                    "enum": [
                                                        "A",
                                                        "B",
                                                        "C",
                                                        "D",
                                                        "F",
                                                        "I",
                                                        "N"
                                                    ]
                                                },
                                                "position": {
                                                    "type": "string",
                                                    "enum": [
                                                        "X",
                                                        "W",
                                                        "H",
                                                        "W*"
                                                    ]
                                                },
                                                "number": {
                                                    "type": "string",
                                                    "example": "25"
                                                },
                                                "amount": {
                                                    "type": "number",
                                                    "example": 5000
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Ticket created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Bets submitted successfully."
                                        },
                                        "data": {
                                            "properties": {
                                                "invoice_number": {
                                                    "type": "string",
                                                    "example": "INV-20260502-0001"
                                                },
                                                "total_amount": {
                                                    "type": "number",
                                                    "example": 15000
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/bets/{id}": {
            "get": {
                "tags": [
                    "Bets"
                ],
                "summary": "Get a ticket with its bets — own only",
                "operationId": "6ae6286b4f92f8ed302fcbd71673babd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ticket detail",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Ticket"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "ErrorResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Unauthorized"
                    },
                    "errors": {
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
            "Ticket": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "invoice_number": {
                        "type": "string",
                        "example": "INV-20260502-0001"
                    },
                    "session": {
                        "type": "string",
                        "enum": [
                            "morning",
                            "noon",
                            "evening"
                        ]
                    },
                    "bet_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2026-05-02"
                    },
                    "total_amount": {
                        "type": "number",
                        "format": "float",
                        "example": 15000
                    },
                    "win_amount": {
                        "type": "number",
                        "format": "float",
                        "example": 0
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "draft",
                            "pending",
                            "settled"
                        ]
                    },
                    "bets_count": {
                        "type": "integer",
                        "example": 3
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}